requirements.txt
requests>=2.31
python-dotenv>=1.0
typer>=0.12
ddgs>=9.0
streamlit>=1.39
pytest>=8.0
